Update script to only remove symbolic links if they exist.

+++ src/clm-pflotran/remove_linked_files.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
 for file in `ls ../pflotran/*.F90`; do
-    rm -f ${file##*/}
+    if [ -e "$file" ]; then
+        rm -f ${file##*/}
+    fi
 done
 
 #for file in `ls ../pflotran/*.F`; do
